ibv-conduit: remove support for XRC w/o PSHM
The support for InfiniBand XRC has always disabled itself at runtime
when the size of a shared-memory domain (supernode) is 1.  Therefore,
XRC support never runs when PSHM is disabled.  So, this commit makes
this dependency explicit at configure time.  This allows removal of
some unreachable code designed to handle the non-PSHM case.