ssh-spawner: continue to optimize SNodeBcast
This commit changes the supernode-scoped broadcast to compute, rather
then communicate, the lengths of each transfer.  Additionally, the
pre-computation of the lengths allows the GatherV data to be received
into a dense buffer regardless of what order it arrives in.  This
means there is no need to repack data before forwarding up the tree in
the GatherV stage, and also simplifies the Indexing step at the root.

Additionally, the gasneti_nodemap[] is used to replace the actual
communication of the 'rootnode' arguments (which, when Gathered,
simply reconstruct gasneti_nodemap[]).  This means that processes
which are not supernode leaders do not participate (not even a 0-byte
send) in the GatherV phase.