pshm: don't cross-map segments in gasnet_init()
This commit resolves bug 3277

With this commit we no longer cross-map segments in gasnet_init().
So, gasnet_attach() now maps them for the first time, rather than
resizing them.

The issue in bug 3277 was that when gasnet_attach() resized segments
to the client-requested sizes, the references due to cross-mapping
could prevent releasing all the resources of the maximum-sized
segments created in gasnet_init(), leading to a temporary over-commit.

On systems with small memory or with tight resource limits (such as
the "shmall" or "shmnmi" parameters for SysV shared memory) the
temporary over-commit was leading to failures as described in bug3277.

Since there is no perceived benefit to having the segments
cross-mapped earlier, the simplest (and most robust) solution was
simply to delay the cross-mapping until gasnet_attach().
Additionally, it is worth noting that this is closer to the way one
expects GASNet-EX to operate.