ssh-spawner: fix error in SNodeBroadcast
For the case of a single process (aka "singleton") and the non-default
N-ary tree topology, gasneti_bootstrapSNodeBroadcast_ssh() was
dereferencing a NULL pointer.  This commit fixes that by guarding the
problematic code with a check for a non-zero value of 'children'.

Note that the code remains structured to favor the non-singleton case.
In other words, the singleton case is rendered *correct* by this
commit, but remains very far from optimal.  If the singleton case is
to be optimized for, then such work should be done at a higher level
so that this function is not even called.