tests: fix code specific to SEGMENT_EVERYTHING
When configured using --enable-segment-everything, the GASNet tests perform an
AM-based all-gather operation at attach time to get a test segment.  Prior to
this commit the code for this all-gather assumed that the seginfo information
for all nodes would fit within a single Medium AM.  This commit adds logic to
loop if necessary to send multiple Mediums if necessary.

Thanks to the Intel PSM team for reporting this issue.