ibv-conduit: facilitate changing GASNETC_BUFSZ
In ibv-conduit the constant GASNETC_BUFSZ sets the size of a
bounce-buffer and indirectly determines AMMaxMedium.  However, there
are at least two other constants which control sizes of particular
Medium messages.  If the value of GASNETC_BUFSZ were to be reduced
without modifying the other two values "bad things" would happen.

This commit ensures that the value of GASNETC_AMRDMA_LIMIT_MAX is
automatically adjusted as GASNETC_BUFSZ changes.

This commit does NOT automate the adjustment of the value of the
GASNETC_PUTINMOVE_LIMIT_MAX constant, but does provide documentation
on how to set it when GASNETC_BUFSZ is changed.