firehose: minor changes for checkpoint support
At least for ibv-conduit we need to deregister all memory before one
can take a checkpoint with BLCR.  This includes, of course, the
registrations which the firehose library is managing.  In the future
I hope to have a graceful way to deal with this, but for now we are
simply calling firehose_fini() before the checkpoint, and then will
call firehose_init() after the checkpoint.

This commit makes some minor changes to firehose_fini() to make it
possible to call firehose_init() after a call to firehose_fini().  As
curently implemented one needs to pass a new flag to firehose_init().
Without that flag we retain the preexisting behavior of returning from
firehose_fini() with the 2 main locks still held (which is intended to
help catch races or coding errors which may lead to calls after fini).