ibv-conduit: compile/link flags for conn thread
Prior to this commit a SEQ build would have the thread flags
(typically -pthread and -lpthread) passed to the compile and link
steps *only* if the AM receive thread was enabled.  This commit
corrects this to pass the thread flags if *either* the AM receive
thread *or* connection progress thread is enabled.

This corrects link errors seen in some non-default configurations.