gasnetrun_ibv.pl: use Cwd::cwd()
This replaces the previous commit (to use Cwd::abs_path) with use of
Cwd::cwd() instead.  This is inteded to ensure that use of ".." and
symlinks are preserved, in case the resolution may not be the same on
the front-end and compute nodes.

Note that this perl script is also the gasnetrun script for (at least)
mxm, ofi and psm.