ssh-spawner: rewrite for BLCR support
This commit is a top-to-bottom rewrite of the ssh-spawner with a
change of architecture to provide support for BLCR, and to generally
make it more adaptible to additional changes.  The design is now
"scalable by default" where the previous default (changable only at
configure time) was to have a single master process with an ssh
connection to each and every rank process.

Prior to this commit the "architecture" of the ssh-spawner was one of
two options, as selected at configure time.  The default option, known
as "flat" featured a single master process on the spawning node, with
each rank process started by a separate ssh command (or by fork() if
local).  This was a poor-scaling design since one file descriptor and
one ssh process was required for each rank process.  The alternative,
known as "Nary" used a tree of ssh connections, and for the common
case would have a single (or perhaps two) ssh processes per compute
node.  However, there was no separate control process on each compute
node and therefore the collective bootstrap communication was
performed *synchronously* within the rank processes.  This lack of any
asynchronous control ability in the "Nary" design made it imposible to
implement some restart scenarios (in particular when a parent needs to
waitpid()).

The new architecture retains the single "master" process on the
spawner node, but now each ssh connection starts a one "control"
process (which is *not* a rank process) per compute node (assuming no
repeats in the node list).  These control processes can optionally be
arranged in a tree for scalability (for instance of file descriptors),
with the choice at runtime (not configure time).  The rank processes
are fork()ed children of the local control process.

The new design makes non-collective operations possible only where
they involve only a rank process and its parent control process.  This
is used, for instance, for partial in-place via BLCR.  However, there
is a possibility of more general operations if changes are made to the
wire protocol (for instance to delimit message boundaries).  While
such changes would be non-trivial, the new design ensures the
necessary relationship among the processes.

There are also several enhancement specificly to support BLCR:
+ Support for command line restarts through both the gasnetrun_*.pl
  script (-restart dir) and explicit master mode (-R dir).
+ Restart even works with "wrapper" commands such as use of env to
  setup the environment:
    gasnetrun_ibv -restart env LD_LIBRARY_PATH=/opt/lib [restart_dir]
  This wrapper support was already available when not performing a
  restart.
+ Support for in-place partial rollback even when BLCR's in-place
  rollback support is not available (which is the case for the public
  release current at this time).
+ Callbacks for integration with the conduit, which must provide for
  any shutdown/restart of the network.