ssh-spawner: two minor fixes
1) This commit corrects the order of the arguments to dup2(), which
were swapped in every single instance.  Calls to dup2() are used only
for two purposes: (a) to redirect STDOUT and STDERR of remote
signal/kill commands to /dev/null, and (b) to redirect STDIN from
/dev/null for all but rank=0.  Since the side-effects of these are
never seen in normal execution, this has gone unnoticed until now.

2) The code to introduce optional delay between ssh commands was
placed *after* fork(), resulting in needless duplication of the
delay in both parent and child.  The delay is now performed only in
the parent, prior to the fork().