ibv-conduit: avoid pthread_cancel() problems
This commit reworks the progress thread machinery to perform
non-blocking calls to ibv_get_cq_event(), moving the blocking instead
to a select() call.  This change is necessary because evidence shows
that pthread_cancel() can interrupt a blocking ibv_get_cq_event() such
that the CQ is left in a state that causes a later ibv_destroy_cq() to
hang (in over 3% of runs in one test I performed).

The only points in the progress loop where cancellation is permitted
are now the new select() call and the optional sleep()s for rate
throttling.  Those are the only blocking calls.

This commit cherry-picked from branch 'feature/blcr'