general: support/use __has_attribute()
In the case that GASNet detects at compile time (of self or client
code) that the compiler is not CC, CXX or MPI_CC from configure time,
we previously had to give up if using gcc >= 4.4.  That was because we
had to assume (conservatively) that __attribute__((__may_alias__)) was
not supported.  However, with gcc-5.x and any recent clang one can
perform *compile-time* checks using __has_attribute().  So, this
commit adds support for use of __has_attribute() as a last-chance
effort when running with a compiler other than the ones probed at
configure time.