aries/gemini conduits: sub-block allocator for AMs
This commit implements an idea suggested by John Bachan for allocating
the remote AM buffers in units less than the maximum AM size.  The AM
buffer space is split into 64 cache-aligned chunks and management of
these chunks replaces credits as the unit of flow control.

The changes are surprisingly minor, with the main changes being:
+ Updates to "remote_request_map" now change multi-bit runs.
+ Some additional (per-target) thread serialization is required to
  prevent starvation of large AMs by small ones in a PAR build.
+ Need buffer allocation logic for contruction of Replies, since one
  can no longer be certain that the buffer used to receive the Request
  is large enough.

Measurments with testam show that any increase in the AM latency (due
to increased injection overhead) is "in the noise".

As a result of the change to how space is allocated, this commit also
raises the AMMaxMedium value to 4k *without* any increase in the per-
peer memory registered for AM buffers.  This yields about a 2.6x
increase in the bandwidth of an AMMedium flood test on Edison or Cori,
and about 2.0x on Titan (between nodes on the same NIC in all cases).