aries/gemini: fix (uncommon) thread perf. issue
This commit swaps the order of "unlock" and "yield" in the busy
waiting loops used when allocating resources for an AM Request.  Prior
to this change, if (and only if) the non-default polite sync mode was
enabled, a busy waiting thread would call sched_yield() while still
holding a lock.  Now the lock is released before the yield, to allow
another thread to potentially obtain it.