aries/gemini: memory reg. caching via UDREG
This commit adds support for use of Cray's UDREG library for caching
of memory registrations.  This leads to a significant improvement in
the bandwidth and latency of out-of-segment Gets, Puts and AM Long
payloads.  This new support is enabled by default.

Some tuning parameters have changed as a result of this new work,
including moving the protocol switch between bounce-buffers and
dynamic registration from 8k down to 4k.