aries/gemini: tune-up AM space allocator
This commit reworks the logic for the recently-implemented sub-block
AM space allocation to ensure the size of the chunks managed are
powers of 2 (no smaller than 64), rather than multiples of 64.  This
allow replacement of integer multiplication and division operations
in the critical paths with corresponding left or right shifts.

Additionally, all values are accepted for GASNET_NETWORKDEPTH_SPACE
and the effective value is adjusted to fit at least 2, but no more
than 64, maximum sized AM Mediums.  Of course, documentation has been
updated to match this new behavior.

Finally, the default Max Medium value has been lowered from 4096 to
4032 to allow (as was the case until recently) three max-sized Mediums
to be in flight at any given time (rather then just 2).