ofi-conduit: remove invalid flexible array member
This commit changes the C99 flexible array member "data" at the end of
struct ofi_am_send_buf to a "normal" array member with length
OFI_AM_MAX_DATA_LENGTH.  This change eliminates the non-standard use
of a struct with a flexible aray member as a member of another struct
(this usage is not permitted by C99 and was being rejected by at least
the Solaris Studio C compiler).

The main observation is that one must use GASNETI_ALIGNUP_NOASSERT(),
rather than GASNETI_ALIGNUP(), when defining OFI_AM_MAX_DATA_LENGTH,
in order to ensure that it is a compile-time constant.