aries/gemini: avoid memcpy of Med Request payloads
This commit modifies the logic for receiving of Medium Requests, and
sending Replies from them, to avoid a memcpy() on the target.  This
has become more important as the AMMaxMedium size has grown larger.

Every AM Reply conveys to the requester the right to reuse the buffer
in which the corresponding Request was received.  If running an AM
Medium Request "in place" this creates a race in which a new Request
may arrive and overwrite the in-use payload.  So, we have been copying
Medium Request payloads to a stack buffer to eliminate the hazard.  We
have taken advantage of this to reuse the buffer as space to assemble
the Reply (well, at least when it fits).

The new logic no longer makes a copy of Medium Request payloads at the
target.  The logic for assembling and sending a Reply is essentially
unchanged for Shorts, Longs, and for Mediums with zero-length payload.
However, if running a Medium Request with non-zero payload there are
two significant changes.  First, the logic to assemble the Reply will
not use the buffer containing the Request payload (since it is still
in use).  Second, the send path will defer the actual post of the
Reply until after the Request handler has returned (to the same point
in the code where an implicit credit will be sent if no Reply has been
issued by the handler).