tests: add more validation in testcore2
This commit adds additional validation to testcore2 in order to detect
if a Medium Request payload is overwritten when sending a Reply.  This
addition is specifically intended to validate recent changes in aries-
and gemini-conduits that could (had they been implemented incorrectly)
lead to such corruption.  However, as a correctness test, this is
generally applicable.