ssh-spawner: fix signal forwarding
In the recent rewrite of ssh-spawner the signal handling behavior
changed subtly such that a termination signal sent to the process
group of the root process was no longer killing the rank processes.
This commit fixes that (mis)behavior by ensuring that any control
process (not just the root) receiving a termination signal will
forward it to its own children.