ibv-conduit: corrections to recent AM code changes
This commit fixes a few minor bugs, and one significant oversight, in
the recent changes to the AM code to allow (for System Requests) the
use of a specific channel.  The code was fundamentally broken for the
case in which SRQ is used (because the QPI offset was applied twice).
Additionally the code continued to use 'dest' where 'node' was
appropriate.