ibv-conduit: no PFs during quiesce
This commit adds logic to help ensure that no progress functions will
run while the quiesce code is sending its "special" AMs.  This is
necessary because in general a PF could require sending an AM Request,
for which we may have already "stolen" all of the necessary credits.

The mechanism for preventing PFs from running is to change the generic
AM Request path to inhibit invoking them for Requests which pass a
handle to block on local completion.  Then the AM Request in the
quiesce code is (artificially) made to pass such a handle (though it
is never checked).  This change is safe (and effectively quite small)
because any System AM that intends to block will run PFs when it does
so.  Additionally, there are currently no other AMs Requests (only
Replies) using the the blocking variant - and thus nothing that might
break due to this change!