autoconf: fix Aries support configure probe
This commit fixes configure to ensure that if the probe for "working
Aries support" were to fail and --enable-aries has been passed, the
configure step will fail.  Due to an instance of "have_gemini" which
was never changed when Aries support was added, the configure step
would previously have run to its conclusion, but one would not have
Aries support despite having passed --enable-aries.