configure: support Aries/Gemini w/o Cray wrapper
This commit adds logic to configure to pull the necessary -I and -L
flags from the enviroment (set by "module load ...") if/when one
configures on a Cray system with a compiler other than Cray's "cc"
wrapper.

The main disadvantage of doing this is that we capture *specific*
versions of xpmem, pmi, ugni and udreg as opposed to allowing the
wrapper compiler to use whatever version is in the environment at
compile/link time.