aries/gemini: lower Aries default UDREG cache size
This commit lowers the default UDREG cache size for Aries systems to
3072 (from 4096) which provides a significant safetly margin relative
to an empirically determined safe value of 3800 or larger.  On Gemini
systems the default remains at the very conservative value of 16.

The README text for the environment variable GASNET_GNI_MEMREG has
also been updated with the new default and with information about the
symptoms of too-large values (error messages from GASNet and MPICH).