ssh-spawner: avoid illegal FD_ISSET(-1, ...)
This commit changes the logic for post-select() scanning of the fdset
to prevent the root process from checking child[-1].  For all control
processes except the root, child[-1] is the parent, but since the
root process has no parent, child[-1].fd has been set to -1.

The problem went unnoticed on Linux, for unknonw reasons.  However, on
Solaris every single ibv-conduit run launched with ssh-spawner was
failing early in job launch.