ibv-conduit: close race on enable/disable of pf
This commit fixes a small race in the recently-added code which moved
the amrdma "balance" work into a progress function.  Prior to this
commit it was possible for a thread to decrement the progress
function's enable counter when it was already zero.  Now the code
looks more like the cannonical example:
  Th0:   write(x); wmb; write(y)
  Th1:   read(y);  rmb; read(x)