Mic/phi support: remove MAP_HUGETLB flag to mmap
As we've discussed in email, the MAP_HUGETLB flag should no longer be
"required" on Xeon-phi systems, as we should be able to rely on
transparent huge page support to do the right think most of the time.
On the other hand, use of this flag can lead to failed munmap() calls
when GASNet tries in gasnet_attach() to "trim" the over-sized segment
created in ganet_init().