Rename cross-configure-intel-mic to ...-knc
There is evidence to suggest that Knights Landing (knl) will be
sufficiently different from knc that we will either need a different
cross-configure script, or perhaps will not want to use one at all.
Either way, this commit renames the cross-configure script for "mic"
to "knc" to reduce the possiblity of confusion in the (near) future.