firehose-region: fix a registration resource leak
This commit corrects a bug in which in certain (rare?) circumstances
the region-based firehose implementation was destroying its record of
a region which was fully covered by other regions without also asking
the conduit to unpin it.  This error has been there ever since the
code to remove covered regions was first added.

It is possible that this explains some rare reports of firehose
pinning failures.  However, there is no good way to confirm that.