handleStdOutput improvements
Rework handleStdOutput to avoid unnecessary extra select() and ioctl(FIONREAD)
calls, and to use a statically-allocated buffer.

Fix a bug this exposed in the call to setupStdSocket, whereby the close of a
listen socket followed by the accept of a new socket aliasing the same socket
id could lead to a spurious invocation of handleStdOutput that could hang under
the new design.