Extend the AMUDP sequence number field
AMUDP sequence number field in the AM header is increased from 1-bit to
6-bits (without changing overall header size).

This fixes Bug 3286 by detecting and dropping UDP packets that arrive
out-of-order (which can happen on systems with multiple network cards or
multiple point-to-point paths in their IP routing tables), that could
previously lead to duplicate or discarded AM handler invocations under
high network load.

A warning message is printed the first time any out-of-order packet
delivery is detected by a worker, since we can't 100% guarantee
correctness in that configuration (although corruption of the AM stream
is now far less likely to arise in practice).

The performance and correctness impact should be neglible for unaffected
configurations, and a significant correctness improvement for affected
systems.