Fix startup envvars for non-local spawning
When AMUDP spawns locally (SPAWNFN=L) the master's spawn environment is
automatically propagated to the forked workers, and everything is
consistent and works correctly.

However when we spawn remotely (SPAWNFN=S) the spawn environment is
propagated explicitly by AMUDP_SPMD partway through the startup
procedure. WORKERIP is special because it's needed so early in startup,
so it's checked on the master and propagated directly on the spawning
command line to the worker.

Unfortunately WORKERIP was incorrectly/redundantly being checked in the
worker node environment (prior to general env propagation).
Consequently, if the worker node happened to have WORKERIP in its
default environment (eg .bashrc), that setting would override the value
sent from the spawning console. It also meant that different workers
(who for example happen to have a different .bashrc) could end up trying
to use different IP subnets for UDP communication in the same job.

This commit removes that problematic second check for WORKERIP on the
worker, which also had the unrelated side-effect of breaking VERBOSEENV
reporting on non-locally spawned workers.

In a related problem, the variables controlling the retransmission
algorithm (REQUESTTIMEOUT_MAX, REQUESTTIMEOUT_INITIAL,
REQUESTTIMEOUT_BACKOFF) were not being correctly honored on non-locally
spawned workers due to consulting the un-propagated environment.

Improve the parsing and defaulting for the retransmission env vars.