Improve SIGPIPE handling
POSIX mandates a SIGPIPE be delivered on a send/write operation to a
disconnected stream socket.  However it's less disruptive to clients
(and potentially more efficient) to handle this possibility via the
return code, which is also guaranteed to be EPIPE in this situation.

Linux and a few others allow the signal delivery to be disabled via the
MSG_NOSIGNAL flag to the send() operation, while others (BSD-based) have
a SO_NOSIGPIPE flag that can be set on the socket via setsockopt().
Others (eg Solaris) lack either mechanism.

We now use SO_NOSIGPIPE and/or MSG_NOSIGNAL where they are available
(#defined), and only otherwise do we mess with the process signal mask
to SIG_IGN potential SIGPIPEs while sending on a stream socket.