fix bug 2428: AMUDP problems on OSX/PPC64
Darwin is apparently expecting a 32-bit argument type to ioctl(FIONREAD), and
we've been passing unsigned long. This worked fine for any ILP32 ABI and happenned
to work correctly for little-endian x86_64, but leads to breakage on 64-bit
big-endian PPC.

Now using unsigned int, which should be correct on all OSX backends.