Reply retransmit optimization
When a node is inattentive for awhile, multiple copies of the same
request can build up in the receive buffer. Previously, this could
result in duplicate back-to-back retransmissions of the same reply. We
now detect this situation by tracking the last approximate transmission
time of a reply in the timestamp status field, and squashing duplicate
copies of the reply (defined as those that would have been re-issued
during the same AMPoll operation).

Never send a reply to an out-of-order request, because it's guaranteed
to be useless.

Remove the transmitCount status field which was never being used.

Use (long long) for stats printf of int64_t on ILP32.