Reduce thread serialization in AMPSHM
To prevent small AMs from starving larger ones, it is necessary to
serialize access to the buffer allocator.  Prior to this commit the
mechanism for this was pthread mutexes (one each for the Request
and Reply channels) held by a thread as it alternated calls to the
per-channel allocator and (a variant of) AMPoll() until a buffer could
be allocated.  However, the "excluded" threads were blocked in
gasneti_mutex_lock() *unable( to Poll for however long it took for the
lock-holding thread to obtain its buffer.

This commit changes the serialization code so that there is a poll
loop in which only one thread may call the allocator, but all threads
will continue to Poll for incoming AMs.  The mutexes still exist, but
the critical section duration is now at most ONE call to the
allocator plus a few statements to manage the serialization; and the
minimum duration is just checking two scalar serialization variables
to determine that another thread "owns" the allocator.

This revised logic will also provide a simple/reliable implementation
of the IMMEDIATE flag in GASNet-EX (just a few additional statements).