AMPSHM: simplify non-threaded buffer allocation
A recent commit to this file added serialization logic intended to
allow concurrent polling while allocating AM buffers.  However, since
that new logic does not rely solely on mutexes, it does not compile
away in a non-threaded build.

This commit "dumbs down" the non-threaded case to entirely omit the
serialization (code and variables).