harness: attempt to build testcxx-{seq,par}
If CXX is configured (via udp-conduit) then the gasnet-tests harness
suite will now attempt to build testcxx-seq and testcxx-par for the
default conduit.  If udp-conduit is not enabled, then we will get a
graceful failure (compile step appears to pass).

This is designed to catch C++ compilance violations in our public
headers.  We already know of one in ibv-conduit/gasnet_core_fwd.h,
which the new tests have reproduced.

There is a good chance that we will encounter link failures that are
entirely unrelated to the test.  So, we may need to devise a
"kill-switch" for these tests if those are not fixable.