atomics: fix signed/unsigned warnings from C++
The recent addition of testcxx to the nightly tests has shown some
minor C++ conformance issues in our public headers (which must be
"clean" to support clients which are written in C++).

This commit fixes instances of passing signed actual parameters where
the function declarations have unsigned formal parameters.  This
includes use of an explicit "-1" in some decrement/dec-and-test
implementations as well as "-(op)" in some subtract implementations.