Fix overridden command Makefile warnings
Teach Bootstrap to enforce selected Makefile rule overrides where
automake is too stupid to notice and suppress the overridden rule
itself.  This silences the spam of make warnings about command
overrides, since we routinely use overrides to workaround the numerous
bugs and deficiencies in automake.

Remove the distclean-recursive hack for automake 1.4 which is no longer
required.

Distclean now verified to work on:
 automake 1.4-p5, 1.7.8, 1.9.6, 1.11.1, 1.13.4, 1.15