Fix bug #3325 - use MPI_Finalized()
This commit adds a defensive check in the mpi-spawner's finalization
path to avoid making MPI calls after an MPI_Finalize() by another
party.

The problem was seen in a hybrid application which had called
MPI_Init() prior to gasnet_init() and later called MPI_Finalize() and
exit(non-zero) on an error-handlng path (not GASNet-aware).  The
atexit/on_exit handler registered by ibv-conduit then attempted to
call MPI_Comm_free(), which resulted in at least one MPI
implementation screaming on stderr and losing the user's exit code.

Thanks to Sean Treichler of the Legion team for reporting the problem
and testing (and correcting) the fix.