Fix an endianness-detection bug on Cygwin
Cygwin's /usr/include/machine/_endian.h (which is pulled in by other
system headers) unconditionally #defines both _BIG_ENDIAN and _LITTLE_ENDIAN,
which was breaking our endianness detection under the right header
include ordering. This never showed up previously because the first/only
include of portable_platform.h was usually quite early, and now we can
re-include the header much later.

Replace the detection of these dubious symbols with the gcc predefined
variants (eg __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__) since these
are more likely to be universal and less likely to conflict with a
system header temporary.

It's possible this weakens our detection logic for non-gcc when we
lack configure information - if that shows up in practice we may
need to re-add a test for the less reliable symbols, conditional on OS.