Rename the undocumented GASNET_ALLOW_OPTIMIZED_DEBUG override
This internal override was never documented or supported, and the GASNet
team feels its existence could be construed as endorsement to mis-use
GASNet's --enable-debug configure mode.

The override will be removed entirely in the next release.

The related AMX overrides of similar name are removed now.

Remove ancient configure workaround for the RampBLUE research
architecture, which never reached production.