ofi: further refine new exit handling code
The major change in this commit is to replace the (recently added) N^2
AMs used for coordination of exits with N*lg(N).  Instead of each
process sending to every other node, we now use the same dissemination
barrier pattern as used in other conduits.  In this pattern each
process now send one AM (and receive one) in each of lg(N) rounds.

In addition to making for a more scalable communications pattern this
change has the effect, due to waiting to receive, of ensuring that we
do not attempt to shutdown libfabric until all AMs have been received.
This eliminates the abortive spawner exits seen for collective normal
client exits (reported as bug 3336).

Additional changes include:
 + The sending and receiving of AMs is now subject to a timeout
   (sampling time between Poll calls) to detect if there are
   inattentive nodes in non-collective exits.
 + If the timeout, above, is reached then the call to
   gasnetc_ofi_exit() is skipped to avoid getting hung as in bug
   3336 (though an abortive spawner exit is still likely to occur).
 + The bootstrapFini() call is now made within an alarm() to catch
   the unlikely possibility that it gets hung.