configure: fix CROSS_HAVE_X86_CMPXCHG16B
Like other CROSS_* variables, CROSS_HAVE_X86_CMPXCHG16B was intended
to be a boolean 0 or 1 value, and that is how it is used in the only
two cross-configure scrips which set it (0 for intel-knc, and 1 for
intel-knl).  However, prior to this commit the configure script was
looking for 'yes' instead of '1'.