psm: fix bug 3340 assertion failures in iop_check
This commit moves the increments of the initiated_{put,get}_cnt fields
of an iop to occur *before* the call to psm2_am_request_short() which
they are accounting for.  This eliminates the race in which the reply
or completion handler for the AM request would increment the
{put,get}_competed_cnt field first (leading to the assertion failures
reported as bug 3340).